[refactor] Promote autodetect_extension to render::framework
